requirements.txt
astroid==1.4.7
backports.functools-lru-cache==1.2.1
blinker==1.4
click==6.6
configparser==3.5.0
coverage==4.1
coveralls==1.1
decorator==4.0.10
docopt==0.6.2
factory-boy==2.7.0
fake-factory==0.5.9
flake8==2.6.2
Flask==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
infinity==1.4
intervals==0.8.0
ipaddress==1.0.16
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.5.0
MySQL-python==1.2.5
passlib==1.6.5
py==1.4.31
pycodestyle==2.0.0
pyflakes==1.2.3
pylint==1.6.1
pytest==2.9.2
pytest-cov==2.3.0
pytest-flake8==0.5
pytest-pylint==0.5.0
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.9
validators==0.10.3
Werkzeug==0.11.10
wrapt==1.10.8
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.0
